+++ pusfume/scripts/mods/pusfume/_pusfume_native_config.lua
@@ -7,6 +7,7 @@ return {
     enabled = false,
     first_person_material_package = false,
     first_person_materials = false,
+    first_person_direct_link = false,
     first_person_unit = false,
     hero_preview_enabled = false,
     fur_child_material = false,